requirements.txt
flask
flask-cors
python-dotenv
pdfplumber
google-genai